mdss: flicker_free: Implement framebuffer state condition
Since the always-on display's brightness application does not work properly with flicker_free's backlight calculation, we now introduce a new condition that will determine whether to only return to mdss the raw value or not depending on framebuffer. This however, assumes that framebuffer does not label *blank as FB_BLANK_UNBLANK when AOD is on.
Also fix indentions in mdss_fb.

Signed-off-by: ederekun <sedrickvince@gmail.com>